LPS-108445 Move EditableDecoration and FragmentContent to their own directory
